Added Check-In screen (Waiting for chapter verdict)
Selecting a table row view in find book creates a new window and passes
the title of the book to this new window (The Check in screen).  The
check in screen consists of the title with a top and bottom border(this
needed to be hacked because appcelerator doesn't support this), and a
text area for the comment.  I am now waiting to see what the verdict is
on the chapter selection before writing that code.